+++ src/tray_icon.h
@@ -9,10 +9,11 @@
 #define WM_TRAYICON (WM_USER + 1)
 #define ID_TRAYICON 1
 
-#define ID_MENU_START     1001
-#define ID_MENU_STOP      1002
-#define ID_MENU_SETTINGS  1003
-#define ID_MENU_EXIT      1004
+#define ID_MENU_START        1001
+#define ID_MENU_STOP         1002
+#define ID_MENU_DEMAND_MODE  1005
+#define ID_MENU_SETTINGS     1003
+#define ID_MENU_EXIT         1004
 
 class TrayIcon {
 public:
@@ -24,6 +25,7 @@ public:
     void updateIcon(bool isStreaming, bool isConnected);
     void showMenu(HWND hWnd);
     void showTooltip(const char* text);
+    void setDemandMode(bool on);
 
     std::function<void()> onStart;
     std::function<void()> onStop;